alien spawning in random position
couple of things happening here:
the game checks for a free slot in the alien array (since there can only be a max of 16)
then it does a random position. if that position is not available it checks again. keeps doing this until it finds an open space. sometimes it takes many attempts, so hopefully this dont cause lag problems. should be fine though; everything runs based on custom ticks, so itll just lag spike and shouldnt affect the player much