hyper alien added
Every 30 ticks (3 seconds), it scurries around for 5 ticks. It bounces off of any walls/aliens it encounters.
Took a bit of debugging because they would just go straight through all other aliens. The issue turned out to just be a spelling error in the for loop, causing it to never run but also never give me an error 😭

Also had to code the basis of alien types to be used for all future aliens as well.